+++ Domain/Contracts/UseCases/Heroes/IGetHeroByIdUseCase.cs
@@ -4,6 +4,6 @@ namespace Domain.Contracts.UseCases.Heroes
 {
     public interface IGetHeroByIdUseCase
     {
-        Task<HeroEntity> GetHeroById(long heroId);
+        Task<HeroEntity> GetHeroById(int heroId);
     }
 }
